Fix MSVC compiler warnings
reg.: conversion from 'size_t' to 'xxx', possible loss of data (C4267)

+++ src/engraving/compat/midi/compatmidirenderinternal.cpp
@@ -361,23 +361,23 @@ static BendPlaybackInfo getBendPlaybackInfo(const GuitarBend* bend, int bendStar
 }
 
 static void fillBendDurations(const Note* bendStartNote, const std::unordered_set<const Note*>& currentNotes,
-                              std::unordered_map<const Note*, size_t>& durations, bool tiedToNext)
+                              std::unordered_map<const Note*, int>& durations, bool tiedToNext)
 {
     if (!bendStartNote || currentNotes.empty()) {
         return;
     }
 
     size_t bendsAmount = tiedToNext ? currentNotes.size() + 1 : currentNotes.size();
-    size_t eachBendDuration = bendStartNote->chord()->actualTicks().ticks() / bendsAmount;
+    int eachBendDuration = bendStartNote->chord()->actualTicks().ticks() / static_cast<int>(bendsAmount);
 
     for (const Note* note : currentNotes) {
         durations.insert({ note, eachBendDuration });
     }
 }
 
-static std::unordered_map<const Note*, size_t> getGraceNoteBendDurations(const Note* note)
+static std::unordered_map<const Note*, int> getGraceNoteBendDurations(const Note* note)
 {
-    std::unordered_map<const Note*, size_t> durations;
+    std::unordered_map<const Note*, int> durations;
     const Note* bendStartNote = nullptr;
     std::unordered_set<const Note*> currentNotes;
